# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IT Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /my-assets:
    get:
      operationId: getMyAssets
      tags: [assets]
      summary: Get assets assigned to current employee
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My assets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetWithDetails"
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: Get all assets
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Available, Assigned, "In Service"]
        - name: department
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Assets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetWithDetails"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetWithDetails"
    put:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetRequest"
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /assign-asset:
    post:
      operationId: assignAsset
      tags: [assignments]
      summary: Assign asset to employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAssetRequest"
      responses:
        "200":
          description: Asset assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments:
    get:
      operationId: getAssignments
      tags: [assignments]
      summary: Get all assignments
      security:
        - bearerAuth: []
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: assetId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Assignments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignmentWithDetails"
  /assignments/{id}/return:
    post:
      operationId: returnAsset
      tags: [assignments]
      summary: Return an asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                remarks:
                  type: string
      responses:
        "200":
          description: Asset returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /acknowledge-asset:
    post:
      operationId: acknowledgeAsset
      tags: [acknowledgments]
      summary: Acknowledge asset receipt
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcknowledgeAssetRequest"
      responses:
        "200":
          description: Acknowledgment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acknowledgment"
  /acknowledgments:
    get:
      operationId: getAcknowledgments
      tags: [acknowledgments]
      summary: Get acknowledgments
      security:
        - bearerAuth: []
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Acknowledgments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcknowledgmentWithDetails"
  /acknowledgments/{id}/pdf:
    get:
      operationId: getAcknowledgmentPdf
      tags: [acknowledgments]
      summary: Download PDF for acknowledgment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, employee]
        isActive:
          type: boolean
        isLocked:
          type: boolean
        failedAttempts:
          type: integer
        lockedAt:
          type: string
          format: date-time
        doj:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - name
        - role
        - isActive
        - isLocked
        - failedAttempts
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        sn:
          type: string
        description:
          type: string
        category:
          type: string
        subCategory:
          type: string
        assetNumber:
          type: string
        owner:
          type: string
        employeeId:
          type: string
        custodian:
          type: string
        department:
          type: string
        email:
          type: string
        assignedDate:
          type: string
        location:
          type: string
        os:
          type: string
        ram:
          type: string
        rom:
          type: string
        osVersion:
          type: string
        serialNumber:
          type: string
        makeModel:
          type: string
        supplier:
          type: string
        invoice:
          type: string
        warranty:
          type: string
        status:
          type: string
          enum: [Available, Assigned, "In Service"]
        remarks:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - assetNumber
        - status
    AssetWithDetails:
      allOf:
        - $ref: "#/components/schemas/Asset"
        - type: object
          properties:
            assignedEmployee:
              $ref: "#/components/schemas/Employee"
            pendingAcknowledgment:
              type: boolean
    CreateAssetRequest:
      type: object
      properties:
        sn:
          type: string
        description:
          type: string
        category:
          type: string
        subCategory:
          type: string
        assetNumber:
          type: string
        owner:
          type: string
        employeeId:
          type: string
        custodian:
          type: string
        department:
          type: string
        email:
          type: string
        assignedDate:
          type: string
        location:
          type: string
        os:
          type: string
        ram:
          type: string
        rom:
          type: string
        osVersion:
          type: string
        serialNumber:
          type: string
        makeModel:
          type: string
        supplier:
          type: string
        invoice:
          type: string
        warranty:
          type: string
        status:
          type: string
          enum: [Available, Assigned, "In Service"]
        remarks:
          type: string
      required:
        - assetNumber
    UpdateAssetRequest:
      type: object
      properties:
        sn:
          type: string
        description:
          type: string
        category:
          type: string
        subCategory:
          type: string
        assetNumber:
          type: string
        owner:
          type: string
        employeeId:
          type: string
        custodian:
          type: string
        department:
          type: string
        email:
          type: string
        assignedDate:
          type: string
        location:
          type: string
        os:
          type: string
        ram:
          type: string
        rom:
          type: string
        osVersion:
          type: string
        serialNumber:
          type: string
        makeModel:
          type: string
        supplier:
          type: string
        invoice:
          type: string
        warranty:
          type: string
        status:
          type: string
          enum: [Available, Assigned, "In Service"]
        remarks:
          type: string
    Assignment:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        employeeId:
          type: integer
        assignedDate:
          type: string
        returnedDate:
          type: string
        remarks:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - assetId
        - employeeId
        - isActive
    AssignmentWithDetails:
      allOf:
        - $ref: "#/components/schemas/Assignment"
        - type: object
          properties:
            asset:
              $ref: "#/components/schemas/Asset"
            employee:
              $ref: "#/components/schemas/Employee"
    AssignAssetRequest:
      type: object
      properties:
        assetId:
          type: integer
        employeeId:
          type: integer
        assignedDate:
          type: string
        remarks:
          type: string
      required:
        - assetId
        - employeeId
    Acknowledgment:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        employeeId:
          type: integer
        assignmentId:
          type: integer
        acknowledgmentDate:
          type: string
          format: date-time
        status:
          type: string
        remarks:
          type: string
        pdfReference:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - assetId
        - employeeId
        - status
    AcknowledgmentWithDetails:
      allOf:
        - $ref: "#/components/schemas/Acknowledgment"
        - type: object
          properties:
            asset:
              $ref: "#/components/schemas/Asset"
            employee:
              $ref: "#/components/schemas/Employee"
    AcknowledgeAssetRequest:
      type: object
      properties:
        assetId:
          type: integer
        assignmentId:
          type: integer
        remarks:
          type: string
      required:
        - assetId
        - assignmentId
...