# lib/api-zod/src/generated/types/acknowledgment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Acknowledgment {
  id: number;
  assetId: number;
  employeeId: number;
  assignmentId?: number;
  acknowledgmentDate?: Date;
  status: string;
  remarks?: string;
  pdfReference?: string;
  createdAt?: Date;
}